requirements.txt
pyQt6
PyMySQL
python-dotenv
fastAPI